Use uint32_t for vertex indices instead of unsigned short. Allows for more vertices in a mesh

+++ include/glade/render/meshes/Cube.h
@@ -118,9 +118,9 @@ namespace Glade {
         return vertices;
       }
 
-      virtual unsigned short* getIndices()
+      virtual uint32_t* getIndices()
       {
-        static unsigned short indices[] = {
+        static uint32_t indices[] = {
           0, 1, 2,
           3, 0, 2,
