+++ include/glade/render/meshes/Rectangle.h
@@ -31,9 +31,9 @@ namespace Glade {
         return vertices;
       }
 
-      virtual unsigned short* getIndices()
+      virtual uint32_t* getIndices()
       {
-        static unsigned short indices[] = { 2, 1, 0, 3, 2, 0 };
+        static uint32_t indices[] = { 2, 1, 0, 3, 2, 0 };
         return indices;
       }
